feat: repeat Encryption works as part of the handler

+++ src/main/java/iaf/ofek/omega/bda/encryptions/CharacterShiftProcess.java
@@ -1,8 +1,10 @@
 package iaf.ofek.omega.bda.encryptions;
 
+import java.math.BigInteger;
+
 @FunctionalInterface
 public interface CharacterShiftProcess {
 
-    String process(Long value, Integer key);
+    String process(BigInteger value, Integer key);
 
 }
