+++ src/main/java/iaf/ofek/omega/bda/encryptions/EncryptionAlgorithm.java
@@ -2,11 +2,13 @@ package iaf.ofek.omega.bda.encryptions;
 
 import iaf.ofek.omega.bda.models.EncryptionKey;
 
+import java.math.BigInteger;
+
 public interface EncryptionAlgorithm<K> {
 
-    String encrypt(Long[] data, EncryptionKey<K> key);
+    String encrypt(BigInteger[] data, EncryptionKey<K> key);
 
-    String decrypt(Long[] data, EncryptionKey<K> key);
+    String decrypt(BigInteger[] data, EncryptionKey<K> key);
 
     EncryptionKey<K> generateEncryptionKey();
