+++ src/main/java/iaf/ofek/omega/bda/encryptions/shift/ShiftEncryption.java
@@ -5,6 +5,8 @@ import iaf.ofek.omega.bda.encryptions.EncryptionAlgorithm;
 import iaf.ofek.omega.bda.models.EncryptionKey;
 import iaf.ofek.omega.bda.utils.RandomUtil;
 
+import java.math.BigInteger;
+
 import static iaf.ofek.omega.bda.consts.EncryptionConstants.CHAR_MAX_VALUE;
 
 public abstract class ShiftEncryption implements EncryptionAlgorithm<Integer> {
@@ -16,12 +18,12 @@ public abstract class ShiftEncryption implements EncryptionAlgorithm<Integer> {
     }
 
     @Override
-    public String encrypt(Long[] data, EncryptionKey<Integer> key) {
+    public String encrypt(BigInteger[] data, EncryptionKey<Integer> key) {
         return processEncryption(data, key.getValue(), this::processCharacterEncryption);
     }
 
     @Override
-    public String decrypt(Long[] data, EncryptionKey<Integer> key) {
+    public String decrypt(BigInteger[] data, EncryptionKey<Integer> key) {
         return processEncryption(data, key.getValue(), this::processCharacterDecryption);
     }
 
@@ -35,16 +37,16 @@ public abstract class ShiftEncryption implements EncryptionAlgorithm<Integer> {
         return new EncryptionKey<>(Integer.parseInt(keyContent));
     }
 
-    protected String processEncryption(Long[] data, Integer key, CharacterShiftProcess characterShiftProcess) {
+    protected String processEncryption(BigInteger[] data, Integer key, CharacterShiftProcess characterShiftProcess) {
         StringBuilder processedData = new StringBuilder();
-        for (Long value : data) {
+        for (BigInteger value : data) {
             processedData.append(characterShiftProcess.process(value, key));
         }
         return processedData.toString();
     }
 
-    protected abstract String processCharacterEncryption(Long value, Integer key);
+    protected abstract String processCharacterEncryption(BigInteger value, Integer key);
 
-    protected abstract String processCharacterDecryption(Long value, Integer key);
+    protected abstract String processCharacterDecryption(BigInteger value, Integer key);
 
 }
